backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary==2.9.6
pillow
drf-spectacular
drf-spectacular-sidecar
django-otp
qrcode
requests
django-cors-headers
gunicorn
